+++ packages/editor/editor-core/src/plugins/text-formatting/commands/clear-formatting.ts
@@ -1,6 +1,5 @@
-import { EditorState, Transaction } from 'prosemirror-state';
 import { liftTarget } from 'prosemirror-transform';
-import { Command } from '../../../commands';
+import { Command } from '../../../types';
 
 export const FORMATTING_NODE_TYPES = [
   'heading',
@@ -19,10 +18,7 @@ export const FORMATTING_MARK_TYPES = [
 ];
 
 export function clearFormatting(): Command {
-  return function(
-    state: EditorState,
-    dispatch: (tr: Transaction) => void,
-  ): boolean {
+  return function(state, dispatch): boolean {
     const { tr } = state;
 
     FORMATTING_MARK_TYPES.forEach(mark => {
@@ -57,7 +53,10 @@ export function clearFormatting(): Command {
     });
 
     tr.setStoredMarks([]);
-    dispatch(tr);
+
+    if (dispatch) {
+      dispatch(tr);
+    }
     return true;
   };
 }