+++ packages/editor/editor-core/src/plugins/type-ahead/commands/items-list-updated.ts
@@ -6,11 +6,13 @@ export const itemsListUpdated = (items: Array<TypeAheadItem>): Command => (
   state,
   dispatch,
 ) => {
-  dispatch(
-    state.tr.setMeta(pluginKey, {
-      action: ACTIONS.ITEMS_LIST_UPDATED,
-      items,
-    }),
-  );
+  if (dispatch) {
+    dispatch(
+      state.tr.setMeta(pluginKey, {
+        action: ACTIONS.ITEMS_LIST_UPDATED,
+        items,
+      }),
+    );
+  }
   return true;
 };