+++ packages/elements/status/src/__tests__/unit/components/internal/color-paletteSpec.tsx
@@ -1,22 +1,23 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import ColorPalette from '../../../components/internal/color-palette';
-import Color from '../../../components/internal/color';
+import ColorPalette from '../../../../components/internal/color-palette';
+import Color from '../../../../components/internal/color';
 
 describe('ColorPalette', () => {
   it('should render 6 colors', () => {
     const onClick = jest.fn();
+    const onHover = jest.fn();
     const component = shallow(
-      <ColorPalette onClick={onClick} selectedColor={'red'} />,
+      <ColorPalette
+        onClick={onClick}
+        onHover={onHover}
+        selectedColor={'red'}
+      />,
     );
-
-    expect(component.find(Color).length).toBe(6);
-    expect(
-      component
-        .find(Color)
-        .first()
-        .props().onClick,
-    ).toBe(onClick);
+    const colorComponent = component.find(Color);
+    expect(colorComponent.length).toBe(6);
+    expect(colorComponent.first().props().onClick).toBe(onClick);
+    expect(colorComponent.first().props().onHover).toBe(onHover);
   });
 
   it('should select selected color', () => {