+++ packages/search/global-search/src/__tests__/unit/common/ContainerSearchResults.test.tsx
@@ -283,7 +283,7 @@ const getPreqQueryResults = (product: Product) =>
     ? getJiraPreqQueryResults()
     : getConfluencePreQueryResults();
 
-['confluence', 'jira'].forEach((product: Product) => {
+(['confluence', 'jira'] as Array<Product>).forEach((product: Product) => {
   describe(`${product} SearchResultsComponent`, () => {
     let searchResultsComponent;
     let getAdvancedSearchUrlSpy;