+++ packages/editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/editor-mobile-bridge
 
+## 6.4.7
+- [patch] [43501db](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/43501db):
+
+  - ED-5812: Fixes some regressions in the mobile editor
+
+  Including:
+   * Disables mediaGoup lazy loading.
+   * Fixes unsupported emoji content.
+   * Fixes missed call to Android bridge for block state.
+
 ## 6.4.6
 - [patch] [e01ea01](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e01ea01):